separando componentes de Testimonial y transfiriendo formato a Nws y Categories
