--- src/Components/Admin/News.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import TableNews from '../Tables/TableNews/TableNews'
-import Fade from 'react-reveal/Fade';
-
-const News = () => {
-	return <Fade>
-    <TableNews/>
-  </Fade>
-}
-
-export default News
