+++ src/Components/Routes/AppRoutes.js
@@ -17,7 +17,7 @@ import ListContainer from '../ListContainer/ListContainer'
 import NewsList from '../News/NewsList'
 import NewsDetail from '../News/NewsDetail'
 import { useSelector } from 'react-redux'
-import TestimonialsList from '../Testimonials/TestiminialsList'
+import TestimonialsList from '../Testimonials/TestimonialsList'
 
 const AppRoutes = () => {
 	const user = useSelector((state) => state.user.currentUser)
